+++ rust_src/src/lib.rs
@@ -91,7 +91,6 @@ mod fileio;
 mod floatfns;
 mod fns;
 mod fonts;
-mod glyphs;
 mod hashtable;
 mod indent;
 mod interactive;
